requirements.txt
pytest==8.4.1
requests==2.32.4
selenium==4.33.0
loguru==0.7.3
python-dotenv==1.1.1
